chore: move mnemonic components into subfolder
