+++ src/app/components/mnemonic/MnemonicInputs/index.tsx
