+++ src/app/components/mnemonic/SecretKeyDescription/index.tsx
